# docs/api/openapi.v1.json
{
  "components": {
    "schemas": {
...
      "BackupActionRequest": {
        "properties": {
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "idempotency_key": {
            "maxLength": 128,
            "minLength": 16,
            "title": "Idempotency Key",
            "type": "string"
          }
        },
        "required": [
          "expected_version",
          "idempotency_key"
        ],
        "title": "BackupActionRequest",
        "type": "object"
      },
      "BackupRequest": {
        "properties": {
          "idempotency_key": {
            "maxLength": 128,
            "minLength": 16,
            "title": "Idempotency Key",
            "type": "string"
          },
          "purpose": {
            "maxLength": 160,
            "minLength": 3,
            "title": "Purpose",
            "type": "string"
          }
        },
        "required": [
          "purpose",
          "idempotency_key"
        ],
        "title": "BackupRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RestoreTestRequest": {
        "properties": {
          "confirmation": {
            "const": "OBNOVIT DO IZOLOVANÉHO PROSTŘEDÍ",
            "title": "Confirmation",
            "type": "string"
          },
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "idempotency_key": {
            "maxLength": 128,
            "minLength": 16,
            "title": "Idempotency Key",
            "type": "string"
          }
        },
        "required": [
          "expected_version",
          "idempotency_key",
          "confirmation"
        ],
        "title": "RestoreTestRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Kanonické API osobní hlasové virtuální asistentky KájovoDagmar.",
    "title": "KájovoDagmar API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/operations/audit": {
      "get": {
        "operationId": "audit_list_api_v1_operations_audit_get",
        "parameters": [
          {
            "in": "query",
            "name": "before_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Before Id"
            }
          },
          {
            "in": "query",
            "name": "from_time",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From Time"
            }
          },
          {
            "in": "query",
            "name": "to_time",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To Time"
            }
          },
          {
            "in": "query",
            "name": "area",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Area"
            }
          },
          {
            "in": "query",
            "name": "result",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 32,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Result"
            }
          },
          {
            "in": "query",
            "name": "actor_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 40,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Actor Type"
            }
          },
          {
            "in": "query",
            "name": "target_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Target Type"
            }
          },
          {
            "in": "query",
            "name": "correlation_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Correlation Id"
            }
          },
          {
            "in": "query",
            "name": "event_name",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Event Name"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "cookie",
            "name": "__Host-kajovodagmar_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "  Host-Kajovodagmar Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Audit List",
        "tags": [
          "operations"
        ]
      }
    },
    "/api/v1/operations/audit/integrity": {
      "get": {
        "operationId": "audit_integrity_api_v1_operations_audit_integrity_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "__Host-kajovodagmar_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "  Host-Kajovodagmar Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Audit Integrity",
        "tags": [
          "operations"
        ]
      }
    },
    "/api/v1/operations/audit/{event_id}": {
      "get": {
        "operationId": "audit_detail_api_v1_operations_audit__event_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "integer"
            }
          },
          {
            "in": "cookie",
            "name": "__Host-kajovodagmar_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "  Host-Kajovodagmar Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Audit Detail",
        "tags": [
          "operations"
        ]
      }
    },
    "/api/v1/operations/backups": {
      "get": {
        "operationId": "backups_api_v1_operations_backups_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "__Host-kajovodagmar_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "  Host-Kajovodagmar Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Backups",
        "tags": [
          "operations"
        ]
      },
      "post": {
        "operationId": "create_backup_api_v1_operations_backups_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "__Host-kajovodagmar_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "  Host-Kajovodagmar Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BackupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Backup",
        "tags": [
          "operations"
        ]
      }
    },
    "/api/v1/operations/backups/{backup_id}/restore-test": {
      "post": {
        "operationId": "restore_test_api_v1_operations_backups__backup_id__restore_test_post",
        "parameters": [
          {
            "in": "path",
            "name": "backup_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Backup Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "__Host-kajovodagmar_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "  Host-Kajovodagmar Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RestoreTestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Restore Test",
        "tags": [
          "operations"
        ]
      }
    },
    "/api/v1/operations/backups/{backup_id}/verify": {
      "post": {
        "operationId": "verify_backup_api_v1_operations_backups__backup_id__verify_post",
        "parameters": [
          {
            "in": "path",
            "name": "backup_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Backup Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "__Host-kajovodagmar_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "  Host-Kajovodagmar Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BackupActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify Backup",
        "tags": [
          "operations"
        ]
      }
    },
    "/api/v1/operations/status": {
      "get": {
        "operationId": "status_api_v1_operations_status_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "__Host-kajovodagmar_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "  Host-Kajovodagmar Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Status",
        "tags": [
          "operations"
        ]
      }
    },
...
  },
...
}